addressing reviewdog findings
