--- catalog-rest-service/src/test/java/org/openmetadata/catalog/selenium/Events.java
@@ -1,14 +1,0 @@
-package org.openmetadata.catalog.selenium;
-
-import org.openqa.selenium.By;
-import org.openqa.selenium.WebDriver;
-import org.openqa.selenium.support.ui.ExpectedConditions;
-import org.openqa.selenium.support.ui.WebDriverWait;
-
-public class Events {
-
-    public static void click(WebDriver driver, By by) {
-        (new WebDriverWait(driver, 30)).until(ExpectedConditions.elementToBeClickable(by));
-        driver.findElement(by).click();
-    }
-}
