+++ catalog-rest-service/src/test/java/org/openmetadata/catalog/selenium/pages/databaseService/DatabaseServicePageTest.java
@@ -17,7 +17,7 @@
 package org.openmetadata.catalog.selenium.pages.databaseService;
 
 import com.github.javafaker.Faker;
-import org.openmetadata.catalog.selenium.Events;
+import org.openmetadata.catalog.selenium.events.Events;
 import org.openmetadata.catalog.selenium.properties.Property;
 import org.openqa.selenium.By;
 import org.openqa.selenium.Keys;
